Improves class inheritance and logic
Class inheritance has been improved so that sub classing is not
redundant. Calls to super() have also been fixed for correct
passing of *args and **kwargs.

The VideoConverter class is now a subclass of ComRemove. This
is done because checking for output file existance occurs in
the VideoConverter.transcode() method. So in the DVRconverter
class, we would try to remove commercials, then check if
transcoded file exists. This wasted a lot of time if the output
file already existed. Now, we can check from file exists, then
run comskip if it does not.

Subtitle grabbing/extracting has also been improved. We now
look for subtitles only after a transcode has completed
successfully. Before it was done first to check if files
should be placed in sub folders, but new convention is to
always place movies in subfolders and never place episodes in
subfolders; they are already in a season folder so redundant.
This improves performance because if transcode fails, time
was not spent getting subtitles for file that no longer exists.

The mediainfo class has been renamed to MediaInfo to follow class
name capitalization convention. The required updates in dependent
code have been made.

v1.0.2