Updates file naming to match plex, better paths
The input naming convetion for movies has been changed so
that the year is included in the title; i.e.,
Movie (YYYY).Extra Info.tt34234134.mkv, or whatever extension.

APPDIR and LOGDIR are variables set in main __init__ file that
specify the "$HOME/Library/Application Support" and
"$APPDIR/Logs", respectively

Still needs to be fully tested, but initial tests are promising.

v1.0.3