Adds new API and classes for metadata
In process of creating new interfaces from TMDb and TVDb APIs so
that they work more like IMDbpy because I like how that is laid
out. For the most part the classes are made, just need to make
some more methods for getting data in to format for mp4Tags and
mkvTags, and some search functionality.

Also, have to update information for input file naming convention
and create function to convert current library file names to
new convention.

v1.2.0