Fixes None bug in os.path.isfile call
In the _cleanUp method of VideoConverter class, arguments
were not check to see if valid (i.e., not None). This caused
exception in call to os.path.isfile. This has been fixed by
checking if arg and os.path.isfile(arg), which will fail before
the os.path.isfile() is reached in the case that arg in None.

v1.2.1