Adds new PopenPool code for better Popen handling
The PopenPool class is ment to act like a multiprocessing.Pool()
object, but for subprocess.Popen(). Basically, you create an
instance, specifying number of threads to allow at one time; may
also specify cpulimit for pool. Then, use the async() method with
same arguments that you would pass to Popen to add a process to
the queue.

v1.2.2