PopenPool.async renamed to Popen_async, bug fixes
The PopenPool.async method has been renamend Popen_async so it
is a little more intuitive as to how it works.

Added ability to set queueDepth for the PopenPool.

Adds try/except in the PopenThread.run() method so that thread
does not break if process does not exits. Also adds a warning
log if non-zero exit status from process.

improves comments.

v1.3.0