Fixes bug with number of threads for processes
New threadCheck utility added to check number of threads requrested
for process. If value input that is NOT integer type, then the
maximum number of threads allowed on machine (multiprocessing.cpu_count())
is returned. Else, if less than 1, return 1; greater than max, return
max, else return input value.

The default value for number of threads across package as been
changed to be maximum number of threads in most cases.

The default number of threads for CLIs are now printed out in the
help (-h, --help). For the watchdogs, it is half of total threads.
for comremove, it is maximum number of threads.

v1.8.4