Adds checks for watchdogs already running
Two (2) new fuctions have been created in the utils subpackage
that use the psutils package to store information about currently
running process and/or check if the stored information matches a
currently running process.

The Plex_DVR_Watchdog and MakeMKV_Watchdog now use the
pidCheck.pidRunning() function right after the argparser is
initialized. The pidRunning function takes a file path as input.
This file is the location of stored process information. If
the file exists, the data are read in and checked against currently
running processes. If the stored information matches the current
information, then it is assumed that an instance of the program is
already running, so the newly started process exits. If no match is
found, the information for the newly started process is stored in
the file.

These changes should eliminate running multiple watchdog instances
and the logging issues those could create.

v1.9.0