Bug fixes in BaseItem and updateFileNames
In the BaseItem() class of the videotagger, there was no private
logger defined, which caused an exception in some cases. This is
fixed by initializing a __log attribute in the __init__ method.

Titles of series, episodes, and movies now have any periods (.)
replaced with under score (_).

In the updateFileNames fuction, now write updated tags to the
newly copied file. Also set the file stats (premissions, created)
to match old file.
Fixes issue with getting the Plex Media Scanner to run correctly.

v1.10.3