Videotagger API bug fixes; filename convert improve
A few bugs in the APIs for TMDb and TVDb have been fixed, these
include:
  - Indexing list of lenght zero (0) in credits parser;
     Now check list has len() > 0
  - Series classes now handle IDs that have tvdb and tmdb in them;
     did not before and would cause 404
  - Replaces periods (.) in episode titles with underscore (_) so
    parsing of info from file name works correctly
  - Fixes issue where missing release/air date would cause error
  - Improves logging for 404 messages
One change/improvement is that & is now replaced with 'and' in
series and episode titles; may still have to upate this for movies.

In the updateFileNames function, many changes have been made in
effort to make more robust. Some of the changes are:
  - Iterate over sub directories in root so that can scan
    individual series/season and movies; saves space.
  - More robust handling of multi-part episodes; i.e., episodes
    that have one (1) IMDb ID but are multiple episodes on TVDb

v1.10.5