Improves file renamer and comments
The file renamer has been improved, again. Now checks to see if
episode returned from byIMDb() search is a Series. If series,
tries to use the TVDbEpisode class to get episode. Then, if not
episode, will fault.

Also fixes bug in the _getRating() method where the country
was not in the release_dates dictionary, causing an exception.
Now use the .get() method to try to get information and test
return from .get() to eliminate faults.

Comments in the videotagger BaseItem class have been improved,
namely to docs for methods.

v1.10.6