Fixes cropping bugs, empty stream issue
In the cropdetect function, there was a 'bug' where videos with
iMax segments (i.e., where go from letter box to full screen)
would end up being cropped. Now, the function checks for the
maximum crop width and height out of all segements, then computes
the crop offsets based on the video size and the crop dimensions.

In some Blu-Rays there were found to be very long segments with no
data in a stream. This would cause ffmpeg to crash as it needs to
know information about all streams. This was 'fixed' by increasing
the max_muxing_queue_size. See https://trac.ffmpeg.org/ticket/6375.

Also added key to set the strict mode for ffmpeg. The default value
is now 'experimental'. This is done to allow some surround tracks
that are not fully supported to be written to MP4 files.

Other changes.
  - Adds get() and videoSize() methods to the MediaInfo class.
  - Increases the time between check for file size changes in the
     MakeMKV watchdog
  - Adds on_moved() method to the MakeMKV watchdog.
  - updateFileNames CLI has been added to main setup.py file.

v1.10.7