Starts implementing logging for transcoder
A rotating log file for video transcoding starts being implemented.
The FFmpegProgress class is added. This class, on instantiation,
holds timing information for updating progress of transcode to
logger. The progress() method is called, passing either a Popen
instance or string object to it for progress updating. If Popen
instance passed, then will read from stdout and process lines,
else, will just process the line.

The RotatingFile class is also added, which acts as a wrapper for
the RotatingFileHandler class. This object can be passed to the
stdout or stderr keywords to subprocess functions/objects for
dumping data to rotating file. Also implements a callback keyword,
which allows dumping lines grabbed from stdout/err to another
function for processing. When using log files for transcoding, the
FFmpegProgress.progess() method is passed in as the callback function
to the RotatinFile class. This enables dumping of ffmpeg logs to a
rotating file AND updating logging with estimated transcode completion
time.

v1.11.0