Changes to user settings and email handlers
The user settings (what used to be the ~/.video_utilsrc) file
has been changes from JSON to YAML format, and is now named
~/.video_utils.yml. This file is copied to location on install
if file matching the name does NOT exist.

The EMailHandler has been updated to parse data from the email
section of the yml file. If all information required for emailing
NOT set in the yml, then will not send emails.

Also adds the new sendEMail decorator function, which is used to
decorate the DVRconverter.convert() method and the new
MakeMKV_Watchdog._process() method. This decorator forces a call
to the EMailHandler.send() method after the decorated methods run.
This enables email updates of processing status.

Note that the EMailHandler will automatically send log info if
a critical log level is encountered.

Other changes:
  - Removes unnecessary dependencies

v1.12.0