Adds TVDb and TMDb attribution, bug fixes
The videotagger now adds attribution to TVDb and TMDb to all
files that are tagged. This information is appended to any
comment written to the file. This is hard-coded and will NOT
be removed.

When using the videotagger CLI, a comment is added saying tags
written using video_utils version x.y.z. When using the
VideoConverter class, comment is added the says video converted
and tagged using video_utils version x.y.z.

Updates/improves comments

Bug fixes:
  - Fixes metadata download issue that arose from setting
    'qualifier' value in self._data in videotagger.BaseItem.
    This caused TMDbMovie (and all other children) to not honor
    the dbid input to the class, so nothing was ever downloaded.
  - Version (i.e., Extended Edition) now specified using version=
    keyword, which sets the self._version attribute in BaseItem
  - Comments and version are now handled correctly when tagging
    files
  - Fixes issue with writing poster to MKV file that already had
    attachments. Now, attempts to delete all attachments before
    writing tags.
  - Adjusts width of version information written to posters so that
    does not run up to poster edge.
  - Check for mkvpropedit will now disable use of mkvTagger function
    instead of raising error; the MKVPROPEDIT set to None if fail
    to import; check MKVPROPEDIT at start of mkvTagger

To do:
  - Continue unit testing
  - Include unit tests in future commits

v1.12.3