Fixes issue with movie version in file rename
In the updateFileName function, changes in setting the item
version in the videotagger subpackage were not correctly
implemented, which lead to no version information being copied
to the renamed file.

To fix this, the videotagger.BaseItem.setVersion() method has
been added and is now used in the updateFileName function.

Also makes the version bar larger on the video posters; is now
5% of poster height instead of 3%

v1.12.4