Moves some label definitions to 'header'
In the videotagger GUI code, the labels and placeholder text for
metadata box entries have been moved to the 'header' of the file
for easier updating in the future.

Also adds a try/except in the videotagger CLI to check for PyQt5
installed. As PyQt5 can have many extra dependencies on systems,
it is not installed on video_utils package install. The try/except
check allows users to still use the CLI to tag files as long as
they do not use the --gui flag. Not fully tested, but should work.

v1.13.4