Improves FFmpeg chapter definition code
Some new classes have been added to the utils/ffmpeg_utils.py
module; namely the FFMetaData and Chapter classes.

The Chapter class is built around the objects returned by ffprobe,
where a Chapter object has a time_base, start and end times, and
a title; more metadata may be added in the future. These objects
can be created by passing in the chapter dictionary from ffprobe,
or by user supplied start and end times, title, and time_base.
This class has a toFFMetaData() method that returns a formatted
string with all information for adding a chapter to a ffmetadata
file.

The FFMetaData class is designed for easy creation of ffmetadata
files. The class has addChapter() and addMetadata() methods for
adding chapters and metadata to files. The addChapter method
accepts either Chapter objects or start and end times, title,
and time_base informaiton. The .save() method is used to generate
the ffmetadata on disc.

The comremove.py module has been updated to use the new Chapter
and FFMetaData classes for cleaner code. The splitOnChapter()
function has also be updated to take advantage of these new classes.

The splitOnChapter() function has also been updated, along with
the addition of some other helper functions, to 'fix' chapter
markers in new video segments. That is to say, chapter times and
titles are updated so that the original chapters 3 and 4 are
updated (both timestamps and names) to be chapters 1 and 2 in the
new segment.

Minimal testing has been performed; only minor, fringe case, bugs
may exist.

v1.14.0