Bug fixes in Chapter class
There were some bugs in the Chapter class, namely, the start_time
could be set to a string, which would cause things to break. Now
only allows setting if value is float type.

In the partialExtract command, removed the -map 0 option from
the ffmpeg command as it was causing outputed files to have
incorrect times.

v1.14.1