Now catches exceptions in main __init__
In the main __init__.py file, excepation was raised if FFmpeg or
mediainfo were not found. This caused documentation building on
RTD to fail as neither of these CLIs exist. Code is now wrapped
in try/except, so user will be able to import the module if those
CLIs do not exists, but things WILL break when trying to use
utilities.

v1.15.3