Updates formatting and imports for RTD
Many utilities, and all code that depends on them, have been
updated to allow building on RTD. The issue was that any code
that requires a CLI would raise an error if the CLI was not
found. While this was handled in higher-level code, when RTD tried
to parse the code for documentation, exceptions were not handled
and things failed.

This version fixes those issues; testing on RTD after commit

v1.15.4