Updates downloadCover to always save file
The videotagger.utils.downloadCover function has been updated
so that it always saves the image to a file on disk. This is done
so that Plex can use the local cover for display because the local
file will have the movie version (i.e., Extended, Uncut, etc.)
banner automatically added.

Improvements:
  - A new function _updateComment() has been added to the writers.py
    file for more robust/consistent appending of TVDb and TMDb
    attribution to file comments.

v1.17.0