Tons of documentation updates!!!
The sphinx documentation has been updated for readthedocs.org.
More needs to be done, namely organizing and updating the front
page, but it is okay for now.

Docstrings in much of the code has been updated so that it
conforms to PEP 257 and 484 (Google style). However, more needs
to be done to finish updating rest of package docstrings.

v1.18.0