Fixes videotagger bugs; docstring updates
In the videotagger subpackage, in the BaseItem class, the __init__
method had data = {} in the definition. This lead to an issue where
the dictionary would be updated to the first data dictionary passed
to it. This ment that any future initialization that did not pass
in a value to the data keyword would use the information from a
different movie, episode, person, etc. The value now defaults to
None and is set in the method.

Also fixes bug in videotagger.parsers.py wherein multiple poster
paths could exist in JSON data, causing the URL to the posters to
have a list in it, which would break. Now check if value is list
or tuple, take first element if it is.

All other changes are changes to docstrings for RTD.

v1.18.3