Fixes bugs in getMetaData() and downloadCover()
The getMetaData() function did not honor the seasonEp keyword
correctly; when a file path was input, this key was ignored. The
key is now correctly used, overriding information from the file
path.

The downloadCover() function had a bug wherein the dump-out return;
i.e., after check that data was downloaded, only returned one (1)
value instead of two (2). This caused the metadata tagging to fail
if poster failed to download. Has been updated so that function has
only one return, with the imagePath set to None by default and
updated if data is downloaded.

v1.19.1