Bug fix for Episode parsing in getMetaData
In the getMetaData method of videotagger.__init__, the regex
search for seaons/episode pattern vailed to grab the first (only)
occurance of the pattern from the result. This ment that the result
was only 1 element instead of 2 (season, episode), causing later
checks to fail. This made the grabber think the file a movie,
leading to more breaking.

v2.0.1