Metadata and file naming bug fixes
Fixes issue with getting metadata from episodes that persisted from
v2.0.1. While that version improved the issue, it did not fully
resolve it in some cases. This is because there was another location
where checked for return from regex matching to have 1 element,
as the regex returns a list. However, the regex result check is
done sooner, so now need to check for 2 elements (season and episode).
This has been updated further down in the code (missed in previous
update) and should fully resolve the issue.

In the videotagger CLI, I failed to replace the --airedOrder flag
with --dvdOrder as the code's default is to search based on aired
order. This has been fixed. Also fixed some logging issues in
the CLI.

For file naming, fixed a 'bug' where a movie or episode title
could have a trailing space before the period delimiter. For
example:

  House (2004)/Season 07/S07E01 - Now What?.tvdbxxxx

This file would have the question mark removed and replaced with
a space, resulting in:

  House (2004)/Season 07/S07E01 - Now What .tvdbxxxx

This has been fixed by running each element of the name through
the replaceChars function, which now strips off leading/trailing
white space using .strip() AND accepts n input arguments. Thus,
the resulting name will now be:

  House (2004)/Season 07/S07E01 - Now What.tvdbxxxx

This goes for movies as well, except for the movie directory IF
there is a year. For example (not real movie):

  Hello? (2004)/

will still become

  Hello  (2004)/

with 2 white spaces, but if no year information, it will be

  Hello/

v2.0.2