Fixes Episode bug in TMDb().byIMDb()
A bug in the byIMDb() method of TMDb() class would cause an
exception because the season and episode number were not passed
to the Episode.TMDbEpisode() object. This has been fixed.

Also fixes a possible bug in the VideoConverter() class where
called metaData.addComment() instead of self.metaData.addComment().
This may not have caused an issue because I believe both reference
the same object, but still, it is cleaner now.

v2.0.3