Fixes bug in getMetaData when no season/ep
There was a large bug in the videotagger.getMetaData function
where when a season/episode was not entered, the function would
crash at the regex search for a season/ep, namely on movies.
This was because the regex.findall() would return and empty list,
which would then be indexed with [0], causing an exception. Have
added a more advance check to see if only one value returned by
the regex. Also added a little bit more debug logging at the end
of the function to print out what metadata was found/downloaded.

Cleaned up some extra semi-colons (;) at the end of lines in the
VideoConverter class.

v2.0.6