Fixes bug in crop detection; no crop detected
There was a bug in the logic for crop detection. When checking
if crop was too close to the original size, was checking
if either the width or height was greater than 95% of the total
resolution. Because the width is rarely cropped at all, this
meant the video was never cropped even if height needed cropped.
Logic has been changed to skip crop if crop width and height
match source, crop if either the crop width or the height is
below the 95% threshold, otherwise, don't crop.

Also move the function responsible for generating the ffmpeg
command used for crop detection to be its own function. Used
to be a sub-function of cropdetect(), but that was silly.

v2.0.7