Adds comdetect flag/keyword to VideoConverter
As Plex now offers non-destructive commercial detection (I know
they have for a while now), I decided it was time to change the
default state of my post processing script so that commercial
detection is DISABLED by default. That way, detection won't be
running twice and we can just let Plex do detection and flag
commercials for skipping.

I have done this by adding a 'comdetect' keyword to the
VideoConverter class. This sets the default state of commercial
detection for the class. When running the transcode method, a
keyword of the same name can be used to override the default
commerical behaviour.

Then in the Plex_DVR_Watchdog script, I have add a --comdetect
flag, which sets the comdetect keyword in the
DVRConverter/VideoConverter class.
Thus, commerical detection is now disabled by default an any
scripts that use Plex_DVR_Watchdog will need to have the
--comskip flag added if you want them to function as they used to.

v2.1.0