Updates TMDb/TVDb tags to match Plex format
With the new Plex Media Scanner, the ability to have TMDb, TVDb,
and IMDb media IDs in the file/directory names has been added to
ensure proper matching every time. Code in the videotagger
subpackage has been updated so that file names follow the Plex
convetion of '{source-ID}' where source is one of tvdb, tmdb, imdb
(lower-case) and the ID is the item number; note that IMDb numbers
begin with 'tt'.

Have also added the rename_media_plex_tag_format CLI to rename
files to match the new Plex convention. Notet that this is a
'dumb' script in that it does NOT run the Plex Media Scanner
while renaming to ensure files remain in the Plex library.
In the other renaming script (updateFileNames), a copy of
every file is created with the new name, then the Plex Media Scanner
is run so that the newly named files are added to the library;
this gives two (2) files for each item, then the orginal is
deleted and another scan is run. That way, the renamed files
do not appear as new files to Plex. However, as above, the
rename_media_plex_tag_format is NOT smart, it does not do all
the extra scanning, and all files MAY show up as newly added.

v2.2.0