Fixes bugs in MakeMKV_Watchdog, add features
There was a bug in the MakeMKV_Watchdog where file size checking
did not check if file existed, which could lead to processing
thread dying. Was also a bug with the _checkSize method not
stopping when an interrupt is encountered; have add a check for
isRunning() to the while loop.

In the updateFilenames utility, added option to exit the program
so user is not forced to do weird things to kill it.

Added beginning stages of a watch dog for adding new music to
plex libraries.

v2.4.0