Uses JSON output of mediainfo
Moves away from using the OLDXML output format from mediainfo in
favor of the JSON formatted output. Had to update some keys in
the MediaInfo class to match those from the JSON format. Also added
2 new functions for parsing int/float from strings and parsing the
JSON output in to a format that is as similar as possible to that
of the OLDXML format