Small bug fixes after last updates
After some testing, realized that I broke some things with last
update.

Fixed issue in MediaInfo where still trying to build self.cmd
attribute; has been commented out.
Fixes 'Bit_depth' key to 'BitDepth' after switch to JSON output
of mediainfo
Fixes missing '-' in the '-profile' flag for ffmpeg in building
options for x265.

Adds back in the OpenSubtiltes class as a base class for the
VideoConverter class as it created some attributes that are
used by subclasses.

Changed TVDb API 'v4' addition back to non 'v4' for
artwork/banners as the 'v4' does not seem to be necesaray.

Addes the 'stream' handler for logging as an actual variable
in the main __init__ file for the package.