Updates Extras support
The support for extra Movie files (e.g., behind the scenes) has
been added. For input file nameing convention, can now use the
-behindthescenes and other tags in what was the 'edition' part
of the input file naming convention to specify what type of
extra the file is. These files are stored in subdirectories for
the movie.

readthedocs have been updated to reflect changes but have not been
fully rebuilt.

Performed minimal linting on code that was being updated.