Linting/updates to Plex_DVR_WatchDog
Linting of the Plex DVR watchdog code.

Also made some updates to improve code flow. Some such updates
are changing the old recording purge thread to a be a thread
timer instead of a thread that sleeps forever. Also improved
some of the purge recordings code logic to reduce nesting of
if statements.