Fixes file naming bug in vobsub convert
In the vobsub_to_srt function, the '.srt' file extension was
incorrectly add to the source file name of the vobsub2srt CLI.
This caused the command to error out as it is expecting the file
name WITHOUT the extension; not to mention that the source file
names for vobsubs end in sub and idx anyway. So, everything was
wrong.

Now have a 'basename' variable that is the subtitle file without
any extension, an 'srtname' that is the basename+'.srt', and
later define the sub_file and idx_file for possible deleting.