Improves Aired/DVD order matching
Noticed that for Arthur (1996) that Plex DVR was naming shows
based on DVD order on TVDb, but the videotagger code was just
grabbing the aired order informaiton for file renaming.

Have added new code in plex.utils that gets metadata for both
aired and DVD order and a new function compare_aired_dvd() that
checks for the metadata search result that has only one (1).
matched. If both have only one (1) match, then we use the
difflib.SequenceMatcher() built-in lib to check if the episode
title from aired order or the dvd order most closely matches the
episode title in the file name.