Improves TVDb episode search logic
Have added better logic to episode search in the TVDb class. When
searching for a specific episode, we now check that the requested
season for the show is with the total number of seasons for each
potential series match. This way, if we are looking for season 10
of a series, and a potential match for the series only has 8
seasons, we do NOT try to look for the episode in the series
because it won't exist. This helps reduce the number of calls to
the TVDb API.