Small bug fix for tvdb2tmdb conversion
When searching for episode of Arthur, a different show
was being found that had an episode with no name/title. This
was causing the tvdb2tmdb() standardization to fail. Added a
try/except when parsing the name and if failes, returns None;
i.e., no match found.

This solution may cause problems down the road.