Updates to videotagger and DVR
The videotagger searching code for the TVDb has been greatly
improved as searches for episodes now try very hard to return
the single best match between aired and dvd order. This is done
by moving some checks done in the Plex DVR renaming code into the
TVDb search code. Now, when search TVDb, if a matching series is
found (based on series title, year, and having a season/episode
number that matches the request), then we look for both aired and
dvd ordered episode. If they return same result (as dvd order falls
back to aired if not dvd ordered ep found), then just give that
episode. Otherwise, we check to see which episode title (aired/dvd)
best matches the requested episode name.

The __eq__ method has also been implemented in the videotagger.BaseItem
class so that comparisons between objects can be made more easily.

In the parsers, I realized that the 'name' key for mulipart episodes
returned by TVDb (think season/series finalies that 'air' as 2
episodes, but are one long episode on DVD) was a list of episode
names. In the parsing code, this would cause the 'title' tag to
be a list while the 'name' tag was updated. This has been changed
so that the 'name' tag retains the original data returned by API
call and the 'title' tag is updated.

For DVR changes, have updated the plex scanner util to accept a
name for the section being scanned; defaults to 'TV Shows'. This
enables the scanning of the 'Movies' section if a movie is DVRed.
Also adds code to extract the top-level directory the DVRed file
is located in ('Movies', 'TV Shows', etc.) to be passed in as the
section name for scanning. Because of this, some changes to the
DVR watchdog had to be made to extract the directory name and then
add it to the DVR queue; a tuple of source file and section name
are now added to the queue instead of just source file.