Bug fixes and RTD updates
Some minor bug fixes in VideoConverter class where creation of
the transcode_postprocess() method led to a variable not being
defined in the new method. Variable is now a 'protected' class
attributed so will exist. Also had changed logic when setting
the output video format based on the is_hdr property and did
things backwards (forgot to check 'not is_hdr' to just 'is_hdr'.

Updates the hdr_utils inject functions so that instead of removing
the source file on successful injection, they rename the new output
file to the input/HEVC file name.

Adds a Read the Docs page for basic over view of how the HDR
pipeline works. Fixes some other bugs in the dependencies file
where a missing link to a website was created in a recent change.