Fixes to Chapter class
Not sure if something changed, but the start_time and end_time
returned by FFProbe for chapter data are stings, which caused
sting values to be returned instead of floats. Have fixed this
by ensuring the start/end setters force values to int with the
properties forced to int on return. The start_time and end_time
have had similar changes applied, but using float()