Modifies audio lang selection, adds fall back
When encoding "Letters from Iwo Jima", there was no english audio track,
so the conversion 'failed'. To prevent this, I modified the audio track
selection by language to check for any tracks that match the requested
language(s) first. If no audio tracks are selected, then there is a log
warning thrown stating that the language of the first audio track will
be used select audio tracks. This is done as the first audio track
should be/is typically the "prominent" audio language for the film.