Changes default number of threads
By default, was using half the number of available threads for
transcoding video. However, I was noticing on new 5950x CPU that for
x264 encodes, using 16 threads over the 6 I had been using on the
5600x was not imporving encoding speed and actually seemed to be
slowing down encodes. So, unless explicity set, the number of threads
to use for encoding should be determined by the encoder, which should
select the optimal number of threads to use.