PGS subs and HVEC fixes
Stumbled on an odd bug that had not hit before with the PGS converter
that caused the pipeline to fail. It seems that the subtitle file
may have been corrupted? Basically could get first image from
subtitles using the pgsripper code, so added a try/except to
catch the issue and just leave the PGS files, as conversion to SRT
failed and would rather have some kind of subtitle file than none
at all.

In the Dolby Vision/HDR10+ extraction code, if both failed, then
dumped out of function before setting the "others_file" attribute,
which is still needed because there is global HDR information for
the film. So, have just moved the defintion of the .mka file to
be just after the outfile so that the pipeline is consistent.