Adds pre/post roll kwargs to splitOnChapter
The split_on_chapter function had one (1) second pre- and post-roll
values for 'fudge' at start and end of segment/chapter splitting.
This could cause some weird things with the start of the next
episode showing up at the very end of the current episode.

While I have NOT changed the default behavior, user can now set
the preroll= or postroll= keywords, or the --preroll or --postroll
CLI flags to adjust this.