Updates for HDR processing
Main update is to handle Dolby Vision FEL data, where a second
video stream is used for difference data in 12 bit streams. This
caused pipeline to fail as was a check for only single video stream.
This has been updated to enforce only single stream for 1080p or
lower resolution and allow for up to two streams for video over
1080, with the assumption that second video stream is FEL data.

In the case of FEL data, then first video stream is re-encoded as
before, with the second stream just copied directly in to output file
in the mkv merge stage.

Have also added new extract_hdr() function that combines the
extraction of dovi and hdr10plus into single command for speed up.
As each of the previous commands had to run ffmpeg for extraction
of video stream, we no just run one extraction and tree that output
into both the dovi_tool and hdr10plus_tool so that both can work
at same time. Is roughly twice as fast as only need single read
of the source file now.

Adds --crop flag to the CLI with default to NOT crop data. As this
is now an option, whereas previous versions enforced cropping, have
added a crop keyword to the new extract_hdr and the dovi_extract
functions.

Renames the hdr_ingect() function to hdr_inject().